Realise that lcio2ascii is not compatible with the modern iLCSoft stack
The lcio2ascii code was written during the time when ILD simulation
studies were based on the Mokka format.
The program now errors out on state-of-the-art simulated files.

We noticed though that the output was mainly based on the LCIO
dumpEventDetailed functionality, which itself is still maintained.
Thus, a lightweight python wrapper should be able to produce the
required output format.
This is left as work for a future commit.